requirements.txt
requests==2.31.0
streamlit==1.26.0
Pillow==9.5.0
pygame==2.5.1